requirements.txt
blinker==1.6.2
cassandra-driver==3.27.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Flask==2.3.2
geomet==0.2.1.post1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
six==1.16.0
Werkzeug==2.3.4
